+++ Studenda.Core/Data/Configuration/MysqlConfiguration.cs
@@ -3,12 +3,12 @@ using Microsoft.EntityFrameworkCore;
 namespace Studenda.Core.Data.Configuration;
 
 /// <summary>
-/// Конфигурация контекста для работы с базами данных MySQL и MariaDB.
+///     Конфигурация контекста для работы с базами данных MySQL и MariaDB.
 /// </summary>
 public class MysqlConfiguration : ContextConfiguration
 {
     /// <summary>
-    /// Конструктор.
+    ///     Конструктор.
     /// </summary>
     /// <param name="connectionString">Строка подключения к базе данных.</param>
     /// <param name="isDebugMode">Статус конфигурации для разработки.</param>
@@ -18,18 +18,18 @@ public class MysqlConfiguration : ContextConfiguration
     }
 
     /// <summary>
-    /// Тип полей даты и времени в базе данных.
+    ///     Тип полей даты и времени в базе данных.
     /// </summary>
     public override string DateTimeType => "DATETIME";
 
     /// <summary>
-    /// Указатель использования текущих даты и времени
-    /// для полей типа <see cref="DateTimeType"/> в базе данных.
+    ///     Указатель использования текущих даты и времени
+    ///     для полей типа <see cref="DateTimeType" /> в базе данных.
     /// </summary>
     public override string DateTimeValueCurrent => "CURRENT_TIMESTAMP";
 
     /// <summary>
-    /// Применить настройки к сессии.
+    ///     Применить настройки к сессии.
     /// </summary>
     /// <param name="optionsBuilder">Набор интерфейсов настройки сессии.</param>
     /// <exception cref="NotImplementedException">По причине отсутствия поддержки.</exception>
