+++ Studenda.Core/Data/DataContext.cs
@@ -2,32 +2,31 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.ChangeTracking;
 using Studenda.Core.Data.Configuration;
 using Studenda.Core.Model;
-using Studenda.Core.Model.Account;
 using Studenda.Core.Model.Common;
-using Studenda.Core.Model.Link;
+using Studenda.Core.Model.Schedule.Management;
+using Studenda.Core.Model.Security;
+using Studenda.Core.Model.Security.Management;
 
 namespace Studenda.Core.Data;
 
 /// <summary>
-/// Сессия работы с базой данных.
-/// 
-/// TODO: Создать класс управления контекстами и их конфигурациями.
-///
-/// Памятка для работы с кешем:
-/// - context.Add() для запроса INSERT.
-///   Объекты вставляются со статусом Added.
-///   При коммите изменений произойдет попытка вставки.
-/// - context.Update() для UPDATE.
-///   Объекты вставляются со статусом Modified.
-///   При коммите изменений произойдет попытка обновления.
-/// - context.Attach() для вставки в кеш.
-///   Объекты вставляются со статусом Unchanged.
-///   При коммите изменений ничего не произойдет.
+///     Сессия работы с базой данных.
+///     TODO: Создать класс управления контекстами и их конфигурациями.
+///     Памятка для работы с кешем:
+///     - context.Add() для запроса INSERT.
+///     Объекты вставляются со статусом Added.
+///     При коммите изменений произойдет попытка вставки.
+///     - context.Update() для UPDATE.
+///     Объекты вставляются со статусом Modified.
+///     При коммите изменений произойдет попытка обновления.
+///     - context.Attach() для вставки в кеш.
+///     Объекты вставляются со статусом Unchanged.
+///     При коммите изменений ничего не произойдет.
 /// </summary>
 public sealed class DataContext : DbContext
 {
     /// <summary>
-    /// Конструктор.
+    ///     Конструктор.
     /// </summary>
     /// <param name="configuration">Конфигурация базы данных.</param>
     public DataContext(ContextConfiguration configuration)
@@ -37,10 +36,7 @@ public sealed class DataContext : DbContext
         // TODO: Использовать асинхронные запросы.
         if (!Database.CanConnect())
         {
-            if (!Database.EnsureCreated())
-            {
-                throw new Exception("Connection error!");
-            }
+            if (!Database.EnsureCreated()) throw new Exception("Connection error!");
         }
         else
         {
@@ -49,58 +45,53 @@ public sealed class DataContext : DbContext
     }
 
     /// <summary>
-    /// Конфигурация базы данных.
+    ///     Конфигурация базы данных.
     /// </summary>
     private ContextConfiguration Configuration { get; }
 
     /// <summary>
-    /// Набор объектов <see cref="User"/>.
+    ///     Набор объектов <see cref="User" />.
     /// </summary>
     public DbSet<User> Users => Set<User>();
 
     /// <summary>
-    /// Набор объектов <see cref="Role"/>.
+    ///     Набор объектов <see cref="Role" />.
     /// </summary>
     public DbSet<Role> Roles => Set<Role>();
 
     /// <summary>
-    /// Набор объектов <see cref="Permission"/>.
+    ///     Набор объектов <see cref="Permission" />.
     /// </summary>
     public DbSet<Permission> Permissions => Set<Permission>();
 
     /// <summary>
-    /// Набор объектов <see cref="Department"/>.
+    ///     Набор объектов <see cref="Department" />.
     /// </summary>
     public DbSet<Department> Departments => Set<Department>();
 
     /// <summary>
-    /// Набор объектов <see cref="Course"/>.
+    ///     Набор объектов <see cref="Course" />.
     /// </summary>
     public DbSet<Course> Courses => Set<Course>();
 
     /// <summary>
-    /// Набор объектов <see cref="Group"/>.
+    ///     Набор объектов <see cref="Group" />.
     /// </summary>
     public DbSet<Group> Groups => Set<Group>();
 
     /// <summary>
-    /// Набор объектов <see cref="WeekType"/>.
+    ///     Набор объектов <see cref="WeekType" />.
     /// </summary>
     public DbSet<WeekType> WeekTypes => Set<WeekType>();
 
     /// <summary>
-    /// Набор объектов <see cref="UserGroupLink"/>.
-    /// </summary>
-    public DbSet<UserGroupLink> UserGroupLinks => Set<UserGroupLink>();
-
-    /// <summary>
-    /// Набор объектов <see cref="RolePermissionLink"/>.
+    ///     Набор объектов <see cref="RolePermissionLink" />.
     /// </summary>
     public DbSet<RolePermissionLink> RolePermissionLinks => Set<RolePermissionLink>();
 
     /// <summary>
-    /// Обработать инициализацию сессии.
-    /// Используется для настройки сессии.
+    ///     Обработать инициализацию сессии.
+    ///     Используется для настройки сессии.
     /// </summary>
     /// <param name="optionsBuilder">Набор интерфейсов настройки сессии.</param>
     protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
@@ -111,8 +102,8 @@ public sealed class DataContext : DbContext
     }
 
     /// <summary>
-    /// Обработать инициализацию модели.
-    /// Используется для дополнительной настройки модели.
+    ///     Обработать инициализацию модели.
+    ///     Используется для дополнительной настройки модели.
     /// </summary>
     /// <param name="modelBuilder">Набор интерфейсов настройки модели.</param>
     protected override void OnModelCreating(ModelBuilder modelBuilder)
@@ -126,15 +117,14 @@ public sealed class DataContext : DbContext
         modelBuilder.ApplyConfiguration(new WeekType.Configuration(Configuration));
 
         // Связующие таблицы.
-        modelBuilder.ApplyConfiguration(new UserGroupLink.Configuration());
-        modelBuilder.ApplyConfiguration(new RolePermissionLink.Configuration());
+        modelBuilder.ApplyConfiguration(new RolePermissionLink.Configuration(Configuration));
 
         base.OnModelCreating(modelBuilder);
     }
 
     /// <summary>
-    /// Сохранить все изменения сессии в базу данных.
-    /// Используется для обновления метаданных модели.
+    ///     Сохранить все изменения сессии в базу данных.
+    ///     Используется для обновления метаданных модели.
     /// </summary>
     /// <returns>Количество затронутых записей.</returns>
     public override int SaveChanges()
@@ -145,13 +135,14 @@ public sealed class DataContext : DbContext
     }
 
     /// <summary>
-    /// Асинхронно сохранить все изменения сессии в базу данных.
-    /// Используется для обновления метаданных модели.
+    ///     Асинхронно сохранить все изменения сессии в базу данных.
+    ///     Используется для обновления метаданных модели.
     /// </summary>
     /// <param name="acceptAllChangesOnSuccess">Флаг принятия всех изменений при успехе операции.</param>
     /// <param name="cancellationToken">Токен отмены.</param>
     /// <returns>Таск, представляющий операцию асинхронного сохранения с количеством затронутых записей.</returns>
-    public override Task<int> SaveChangesAsync(bool acceptAllChangesOnSuccess, CancellationToken cancellationToken = default)
+    public override Task<int> SaveChangesAsync(bool acceptAllChangesOnSuccess,
+        CancellationToken cancellationToken = default)
     {
         UpdateTrackedEntityMetadata();
 
@@ -159,13 +150,13 @@ public sealed class DataContext : DbContext
     }
 
     /// <summary>
-    /// Обновить метаданные всех модифицироанных моделей
-    /// в кеше сессии.
-    /// Такой подход накладывает дополнительные ограничения
-    /// при работе с сессиями. Необходимо учитывать, что
-    /// для обновления записей нужно сперва загрузить эти
-    /// записи в кеш сессии, чтобы трекер корректно
-    /// зафиксировал изменения.
+    ///     Обновить метаданные всех модифицироанных моделей
+    ///     в кеше сессии.
+    ///     Такой подход накладывает дополнительные ограничения
+    ///     при работе с сессиями. Необходимо учитывать, что
+    ///     для обновления записей нужно сперва загрузить эти
+    ///     записи в кеш сессии, чтобы трекер корректно
+    ///     зафиксировал изменения.
     /// </summary>
     private void UpdateTrackedEntityMetadata()
     {
@@ -173,10 +164,7 @@ public sealed class DataContext : DbContext
 
         foreach (var entry in entries)
         {
-            if (entry.Entity is not Entity entity)
-            {
-                continue;
-            }
+            if (entry.Entity is not Entity entity) continue;
 
             // Текущая дата и время на устройстве.
             // Нельзя допустить, чтобы эти данные передавались во внешние хранилища.
@@ -185,7 +173,7 @@ public sealed class DataContext : DbContext
     }
 
     /// <summary>
-    /// Определить, является ли вхождение трекера обновленной моделью.
+    ///     Определить, является ли вхождение трекера обновленной моделью.
     /// </summary>
     /// <param name="entry">Вхождение из трекера изменений.</param>
     /// <returns>Статус проверки.</returns>
