+++ Studenda.Core/Model/Common/Department.cs
@@ -1,42 +1,14 @@
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 using Studenda.Core.Data.Configuration;
+using Studenda.Core.Model.Security;
 
 namespace Studenda.Core.Model.Common;
 
 /// <summary>
-/// Факультет.
+///     Факультет.
 /// </summary>
-public class Department : Entity
+public class Department : Identity
 {
-    /// <summary>
-    /// Конфигурация модели <see cref="Department"/>.
-    /// </summary>
-    internal class Configuration : Configuration<Department>
-    {
-        /// <summary>
-        /// Конструктор.
-        /// </summary>
-        /// <param name="configuration">Конфигурация базы данных.</param>
-        public Configuration(ContextConfiguration configuration) : base(configuration) { }
-
-        /// <summary>
-        /// Задать конфигурацию для модели.
-        /// </summary>
-        /// <param name="builder">Набор интерфейсов настройки модели.</param>
-        public override void Configure(EntityTypeBuilder<Department> builder)
-        {
-            builder.Property(department => department.Name)
-                .HasMaxLength(NameLengthMax)
-                .IsRequired(IsNameRequired);
-
-            builder.HasMany(department => department.Courses)
-                .WithOne(course => course.Department)
-                .HasForeignKey(course => course.DepartmentId);
-
-            base.Configure(builder);
-        }
-    }
-
     /*                   __ _                       _   _
      *   ___ ___  _ __  / _(_) __ _ _   _ _ __ __ _| |_(_) ___  _ __
      *  / __/ _ \| '_ \| |_| |/ _` | | | | '__/ _` | __| |/ _ \| '_ \
@@ -46,17 +18,54 @@ public class Department : Entity
      * Константы, задающие базовые конфигурации полей
      * и ограничения модели.
      */
+
     #region Configuration
 
     /// <summary>
-    /// Максимальная длина поля <see cref="Name"/>.
+    ///     Максимальная длина поля <see cref="Name" />.
     /// </summary>
     public const int NameLengthMax = 128;
 
     /// <summary>
-    /// Статус необходимости наличия значения в поле <see cref="Name"/>.
+    ///     Статус необходимости наличия значения в поле <see cref="Name" />.
     /// </summary>
     public const bool IsNameRequired = true;
+
+    /// <summary>
+    ///     Конфигурация модели <see cref="Department" />.
+    /// </summary>
+    internal class Configuration : Configuration<Department>
+    {
+        /// <summary>
+        ///     Конструктор.
+        /// </summary>
+        /// <param name="configuration">Конфигурация базы данных.</param>
+        public Configuration(ContextConfiguration configuration) : base(configuration)
+        {
+            // PASS.
+        }
+
+        /// <summary>
+        ///     Задать конфигурацию для модели.
+        /// </summary>
+        /// <param name="builder">Набор интерфейсов настройки модели.</param>
+        public override void Configure(EntityTypeBuilder<Department> builder)
+        {
+            builder.Property(department => department.Name)
+                .HasMaxLength(NameLengthMax)
+                .IsRequired();
+
+            builder.HasMany(department => department.Groups)
+                .WithOne(group => group.Department)
+                .HasForeignKey(group => group.DepartmentId);
+
+            builder.HasMany(department => department.Users)
+                .WithOne(user => user.Department)
+                .HasForeignKey(course => course.DepartmentId);
+
+            base.Configure(builder);
+        }
+    }
 
     #endregion
 
@@ -69,17 +78,23 @@ public class Department : Entity
      * Поля данных, соответствующие таковым в таблице
      * модели в базе данных.
      */
+
     #region Entity
 
     /// <summary>
-    /// Название.
+    ///     Название.
     /// </summary>
     public string Name { get; set; } = null!;
 
     #endregion
 
     /// <summary>
-    /// Связанные объекты <see cref="Course"/>.
+    ///     Связанные объекты <see cref="Group" />.
     /// </summary>
-    public List<Course> Courses { get; set; } = null!;
+    public List<Group> Groups { get; set; } = null!;
+
+    /// <summary>
+    ///     Связанные объекты <see cref="User" />.
+    /// </summary>
+    public List<User> Users { get; set; } = null!;
 }
