+++ Studenda.Core/Model/Identity.cs
@@ -1,0 +1,12 @@
+namespace Studenda.Core.Model;
+
+/// <summary>
+///     Модель <see cref="Entity" /> с уникальным идентификатором.
+/// </summary>
+public class Identity : Entity
+{
+    /// <summary>
+    ///     Идентификатор.
+    /// </summary>
+    public int Id { get; set; }
+}
