+++ Studenda.Core/Model/Security/Management/Permission.cs
@@ -1,13 +1,13 @@
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 using Studenda.Core.Data.Configuration;
-using Studenda.Core.Model.Security;
 
-namespace Studenda.Core.Model.Common;
+namespace Studenda.Core.Model.Security.Management;
 
 /// <summary>
-///     Группа.
+///     Разрешение для <see cref="Role" />.
+///     Обозначает некоторый доступ к некоторой функции.
 /// </summary>
-public class Group : Identity
+public class Permission : Identity
 {
     /*                   __ _                       _   _
      *   ___ ___  _ __  / _(_) __ _ _   _ _ __ __ _| |_(_) ___  _ __
@@ -27,24 +27,14 @@ public class Group : Identity
     public const int NameLengthMax = 128;
 
     /// <summary>
-    ///     Статус необходимости наличия значения в поле <see cref="CourseId" />.
-    /// </summary>
-    public const bool IsCourseIdRequired = true;
-
-    /// <summary>
-    ///     Статус необходимости наличия значения в поле <see cref="DepartmentId" />.
-    /// </summary>
-    public const bool IsDepartmentIdRequired = true;
-
-    /// <summary>
     ///     Статус необходимости наличия значения в поле <see cref="Name" />.
     /// </summary>
     public const bool IsNameRequired = true;
 
     /// <summary>
-    ///     Конфигурация модели <see cref="Group" />.
+    ///     Конфигурация модели <see cref="Permission" />.
     /// </summary>
-    internal class Configuration : Configuration<Group>
+    internal class Configuration : Configuration<Permission>
     {
         /// <summary>
         ///     Конструктор.
@@ -59,25 +49,15 @@ public class Group : Identity
         ///     Задать конфигурацию для модели.
         /// </summary>
         /// <param name="builder">Набор интерфейсов настройки модели.</param>
-        public override void Configure(EntityTypeBuilder<Group> builder)
+        public override void Configure(EntityTypeBuilder<Permission> builder)
         {
-            builder.Property(group => group.Name)
+            builder.Property(permission => permission.Name)
                 .HasMaxLength(NameLengthMax)
                 .IsRequired();
 
-            builder.HasOne(group => group.Course)
-                .WithMany(course => course.Groups)
-                .HasForeignKey(group => group.CourseId)
-                .IsRequired();
-
-            builder.HasOne(group => group.Department)
-                .WithMany(course => course.Groups)
-                .HasForeignKey(group => group.DepartmentId)
-                .IsRequired();
-
-            builder.HasMany(group => group.Users)
-                .WithOne(user => user.Group)
-                .HasForeignKey(course => course.GroupId);
+            builder.HasMany(permission => permission.RolePermissionLinks)
+                .WithOne(link => link.Permission)
+                .HasForeignKey(link => link.PermissionId);
 
             base.Configure(builder);
         }
@@ -98,16 +78,6 @@ public class Group : Identity
     #region Entity
 
     /// <summary>
-    ///     Идентификатор связанного объекта <see cref="Common.Course" />.
-    /// </summary>
-    public int CourseId { get; set; }
-
-    /// <summary>
-    ///     Идентификатор связанного объекта <see cref="Common.Department" />.
-    /// </summary>
-    public int DepartmentId { get; set; }
-
-    /// <summary>
     ///     Название.
     /// </summary>
     public string Name { get; set; } = null!;
@@ -115,17 +85,7 @@ public class Group : Identity
     #endregion
 
     /// <summary>
-    ///     Связанный объект <see cref="Common.Course" />.
+    ///     Связанные объекты <see cref="RolePermissionLink" />.
     /// </summary>
-    public Course Course { get; set; } = null!;
-
-    /// <summary>
-    ///     Связанный объект <see cref="Common.Department" />.
-    /// </summary>
-    public Department Department { get; set; } = null!;
-
-    /// <summary>
-    ///     Связанные объекты <see cref="User" />.
-    /// </summary>
-    public List<User> Users { get; set; } = null!;
+    public List<RolePermissionLink> RolePermissionLinks { get; set; } = null!;
 }
